Ecommerce_BackEnd/vaishaliGold/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.42.2
Django==5.1.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-rest-passwordreset==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
google-auth==2.38.0
idna==3.10
pillow==11.1.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
